kill aggUtils
